kludge to get this working with gitlab
Gitlab doesn't do the hash thing, so one can get this to work
by taking the generated hash string, and pasting that in as the
secret in gitlab, so that it gets sent over the wire and then matches
the result of generating the hash at this end.

It would clearly make more sense to do the hash step conditionally,
only when verifying the GitHub hash, and otherwise use the plain text.